+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.10.1] / 2025-05-13
+### Features
+- Support `disable` tests using `NUnit.Application` configuration. (Fix: #73)
+### TestAdapter
+- Add `IsApplicationDisable` to disable tests in `TestDiscoverer`.
+
 ## [1.10.0] / 2025-04-03
 ### Features
 - Support `dotnet test` with `--filter` with properties `FullyQualifiedName` and `Name`. (Fix: #76)
@@ -603,6 +609,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - [x] TestsFail
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.10.1]: ../../compare/1.10.0...1.10.1
 [1.10.0]: ../../compare/1.9.1...1.10.0
 [1.9.1]: ../../compare/1.9.0...1.9.1
 [1.9.0]: ../../compare/1.8.0...1.9.0